moved constant,structs to common file
